api/requirements.txt
fastapi==0.109.2
pydantic==2.6.1
uvicorn==0.27.1
supabase==2.3.4
PyJWT==2.8.0
